separate the main loop out
