+++ validator/core/weight_setting.py
